refactor: move 'flattened' types from subgraph crate to common crate, rm unncessary custom error type
