+++ crates/common/Cargo.toml
@@ -34,6 +34,8 @@ serde_bytes = { workspace = true }
 forker = { workspace = true }
 tokio = { workspace = true }
 typeshare = { workspace = true }
+csv = { workspace = true }
+chrono = { workspace = true }
 
 [dev-dependencies]
 tokio = { workspace = true }
