+++ crates/common/src/types/mod.rs
@@ -1,3 +1,13 @@
+mod order_clears_list_flattened;
 mod order_detail_extended;
+mod order_takes_list_flattened;
+mod orders_list_flattened;
+mod token_vault_flattened;
+mod vault_balance_change_flattened;
 
+pub use order_clears_list_flattened::*;
 pub use order_detail_extended::*;
+pub use order_takes_list_flattened::*;
+pub use orders_list_flattened::*;
+pub use token_vault_flattened::*;
+pub use vault_balance_change_flattened::*;
