+++ crates/subgraph/Cargo.toml
@@ -16,7 +16,6 @@ typeshare = { workspace = true }
 serde = { workspace = true }
 alloy-primitives = { workspace = true }
 rain_orderbook_bindings = { workspace = true }
-csv = { workspace = true }
 chrono = { workspace = true }
 
 [dev-dependencies]
