+++ crates/subgraph/src/lib.rs
@@ -1,4 +1,3 @@
-mod csv;
 mod cynic_client;
 mod orderbook_client;
 mod pagination;
@@ -9,6 +8,5 @@ mod vault_balance_changes_query;
 #[cynic::schema("orderbook")]
 pub mod schema {}
 
-pub use csv::{TryIntoCsv, TryIntoCsvError};
 pub use orderbook_client::{OrderbookSubgraphClient, OrderbookSubgraphClientError};
 pub use pagination::{PageQueryClient, PaginationArgs};
