+++ crates/subgraph/src/types/mod.rs
@@ -1,4 +1,3 @@
-pub mod flattened;
 pub mod order_clears_list;
 pub mod order_detail;
 pub mod order_detail_traits;
@@ -9,3 +8,5 @@ pub mod vault_balance_change;
 pub mod vault_balance_changes_list;
 pub mod vault_detail;
 pub mod vaults_list;
+
+pub use cynic::Id;
