+++ crates/subgraph/src/utils/mod.rs
@@ -1,5 +1,3 @@
 mod slice_list;
-mod timestamp;
 
 pub use slice_list::*;
-pub use timestamp::*;
