+++ tauri-app/src-tauri/src/commands/order_take.rs
@@ -1,9 +1,10 @@
 use crate::error::CommandResult;
-use rain_orderbook_common::subgraph::SubgraphArgs;
+use rain_orderbook_common::{
+    types::OrderTakeFlattened, utils::timestamp::FormatTimestampDisplayError,
+    subgraph::SubgraphArgs, csv::TryIntoCsv,
+};
 use rain_orderbook_subgraph_client::{
-    types::{flattened::{OrderTakeFlattened, TryIntoFlattenedError}, order_takes_list},
-    TryIntoCsv,
-    PaginationArgs,
+    types::order_takes_list, PaginationArgs,
 };
 use std::path::PathBuf;
 use std::fs;
@@ -36,7 +37,7 @@ pub async fn order_takes_list_write_csv(
     let order_takes_flattened: Vec<OrderTakeFlattened> = order_takes
             .into_iter()
             .map(|o| o.try_into())
-            .collect::<Result<Vec<OrderTakeFlattened>, TryIntoFlattenedError>>()?;
+            .collect::<Result<Vec<OrderTakeFlattened>, FormatTimestampDisplayError>>()?;
     let csv_text = order_takes_flattened.try_into_csv()?;
     fs::write(path, csv_text)?;
